Started Work on Charts
Added in the ChartWindow, and BasicChart, and set up a new opengl rendering system for the charts that deals strictly with 2d lines. The render system for the map will likely be very similar in design, but much more complex. Currently, only the framework for rendering lines and the util functions for determining grid divisions are in place. Line limiting on the input box was added in, and the logic loop for all game components was implemented for future use in the tick() functions.