Made a Chart Demo
A new window was created to test input boxes and the new chart system (most functionality was added this commit). BasicChart can graph 2 or more points and connect them with lines, properly displaying axis divisions and labels, and of course everything scales with window size and GUIScale. A test window was made with the ability to manually enter any number of points of x and y to graph on a chart, with a test button for a sine wave. Improvements were made to focus events on TypeBoxes and buttons.