Map Implemented
MapBase is now fully functional, rendering planets, stars, and orbit circles for a solar system. Also improved the chart rendering, using GL_LINE_STRIP instead of GL_LINE to reduce the number of vertices needed by half. Temporary textures were added for a sun and planet to be rendered on the map, though custom rendering is likely desired in the near future.