Reflective, Recursive Property Windows
Implemented a Property window that auto-generates the labels and updatable values for all public variables labeled with custom tags. Improvements to the general design of the window should be easy, as the system only populates the needed text boxes at compiletime. VSync was also fixed (again), and slight adjustments were made to the TypeBoxes for functional stability.